move cli examples to examples/cli
