+++ trubrics/cli/main.py
@@ -18,14 +18,13 @@ app = typer.Typer()
 def run(
     save_ui: bool = False,
     trubric_config_path: str = typer.Option(
-        ".",
+        ...,
         prompt=(
-            "Enter the path to your trubric config .json file (this file can be generated by running `trubrics init`)."
-            " The default path is"
+            "Enter the path to your trubric config .json file (this file can be generated by running `trubrics init`)"
         ),
     ),
     trubric_output_file_path: str = typer.Option(
-        ".", prompt="Enter the path to save your output trubric file. The default path is"
+        ".", prompt="Enter a path to save your output trubric file. The default path is"
     ),
     trubric_output_file_name: str = typer.Option(
         "my_new_trubric.json", prompt="Enter the file name of your output trubric file. The default file name is"
@@ -123,11 +122,11 @@ def run(
 
 @app.command()
 def init(
-    trubric_config_path: str = typer.Option(
-        ".", prompt="Enter the path to your trubric config .json file. The default path is"
-    ),
     trubric_run_path: str = typer.Option(
-        ..., prompt="Enter the path to your trubric run .py file (e.g. examples/trubric_run.py)"
+        ..., prompt="Enter the path to your trubric run .py file (e.g. examples/cli/trubric_run.py)"
+    ),
+    trubric_config_path: str = typer.Option(
+        ".", prompt="Enter a path to save your .trubrics_config.json. The default path is"
     ),
 ):
     """The CLI `trubrics init` command for initialising trubrics config."""
